safari-600.7-branch/LayoutTests/platform/ios-sim/fast/events/scale-and-scroll-body-expected.txt

FAIL window.document.body.scrollTop should be 30. Was 0.
FAIL window.document.body.scrollLeft should be 30. Was 0.
PASS window.document.body.scrollTop is 100
FAIL window.document.body.scrollLeft should be 100. Was 0.
FAIL window.scrollX should be 100. Was 0.
PASS window.scrollY is 100
